Split Store and TreeObservable
