+++ Source/Fuse.Reactive.JavaScript/Fuse.Reactive.JavaScript.unoproj
@@ -39,7 +39,8 @@
     "3rdparty/es6-promise.min.js:File",
     "FuseJS/setTimeout.js:File",
     "FuseJS/Observable.js:File",
-    "FuseJS/TreeObservable.js:File",
+    "FuseJS/TreeObservable.js:Bundle",
+    "FuseJS/Store.js:Bundle",
     "FuseJS/EventEmitter.js:Bundle",
     "JavaScriptEngine.cil.uxl:Extensions",
     "JavaScriptEngine.cpp.uxl:Extensions",
